Add pointer shortcuts to help section.
